Return the envelope every time
I didn't like how complicated the base request function had become, so I've updated it to only return the envelope instead of giving you the option to return the data map, the envelope, or the entire response. You can easily obtain the data map from the envelope using the :data key. This change removes the :return option.

Being able to get the entire response is still useful for debugging purposes, though. For this I added a :debug? option that if true will attach the entire HTTP response map as metadata to the envelope.